CORE-489 Increase size of some variables(int->size_t) in ConnectBlock and VerifyDB

Approved-by: Richard Mills <r.mills@nchain.com>

+++ src/validation.cpp
@@ -3388,7 +3388,7 @@ static bool ConnectBlock(
 
     std::vector<int> prevheights;
     Amount nFees(0);
-    int nInputs = 0;
+    size_t nInputs = 0;
 
     // Sigops counting. We need to do it again because of P2SH.
     uint64_t nSigOpsCount = 0;
@@ -3578,7 +3578,7 @@ static bool ConnectBlock(
 
     nTimeVerify += nTime4 - nTime2;
     LogPrint(BCLog::BENCH,
-             "    - Verify %u txins: %.2fms (%.3fms/txin) [%.2fs]\n",
+             "    - Verify %zu txins: %.2fms (%.3fms/txin) [%.2fs]\n",
              nInputs - 1, 0.001 * (nTime4 - nTime2),
              nInputs <= 1 ? 0 : 0.001 * (nTime4 - nTime2) / (nInputs - 1),
              nTimeVerify * 0.000001);
@@ -6139,7 +6139,7 @@ bool CVerifyDB::VerifyDB(const Config &config, CCoinsView *coinsview,
     CCoinsViewCache coins(coinsview);
     CBlockIndex *pindexState = chainActive.Tip();
     CBlockIndex *pindexFailure = nullptr;
-    int nGoodTransactions = 0;
+    size_t nGoodTransactions = 0;
     CValidationState state;
     int reportDone = 0;
     LogPrintf("[0%%]...");
@@ -6231,7 +6231,7 @@ bool CVerifyDB::VerifyDB(const Config &config, CCoinsView *coinsview,
 
     if (pindexFailure) {
         return error("VerifyDB(): *** coin database inconsistencies found "
-                     "(last %i blocks, %i good transactions before that)\n",
+                     "(last %i blocks, %zu good transactions before that)\n",
                      chainActive.Height() - pindexFailure->nHeight + 1,
                      nGoodTransactions);
     }
@@ -6265,7 +6265,7 @@ bool CVerifyDB::VerifyDB(const Config &config, CCoinsView *coinsview,
     }
 
     LogPrintf("[DONE].\n");
-    LogPrintf("No coin database inconsistencies in last %i blocks (%i "
+    LogPrintf("No coin database inconsistencies in last %i blocks (%zu "
               "transactions)\n",
               chainActive.Height() - pindexState->nHeight, nGoodTransactions);
